Split Launcher Services into domain folders
Group the 23-file Services folder into Install/ (install pipeline and its
exceptions), Feeds/ (release feeds and self-update), News/, and
Infrastructure/ (logging, dialogs, config, window plumbing), with
namespaces matching the folders and consumers updated.
