+++ Ksp2Redux.Tools.Launcher.Test/CacheServiceTest.cs
@@ -1,6 +1,6 @@
 using System.IO.Compression;
 using Ksp2Redux.Tools.Common.Services;
-using Ksp2Redux.Tools.Launcher.Services;
+using Ksp2Redux.Tools.Launcher.Services.Install;
 using Moq;
 using Testably.Abstractions.Testing;
