+++ Ksp2Redux.Tools.Launcher.Test/CommunityTabViewModelTest.cs
@@ -1,5 +1,6 @@
 using Ksp2Redux.Tools.Launcher.Models;
-using Ksp2Redux.Tools.Launcher.Services;
+using Ksp2Redux.Tools.Launcher.Services.News;
+using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 using Ksp2Redux.Tools.Launcher.ViewModels.Community;
 using Moq;
