+++ Ksp2Redux.Tools.Launcher.Test/HeadlessTests/TestAppBuilder.cs
@@ -2,7 +2,10 @@
 using Avalonia;
 using Avalonia.Headless;
 using Ksp2Redux.Tools.Common.Services;
-using Ksp2Redux.Tools.Launcher.Services;
+using Ksp2Redux.Tools.Launcher.Services.Install;
+using Ksp2Redux.Tools.Launcher.Services.Feeds;
+using Ksp2Redux.Tools.Launcher.Services.News;
+using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 using Ksp2Redux.Tools.Launcher.ViewModels;
 using Ksp2Redux.Tools.Launcher.ViewModels.Community;
 using Ksp2Redux.Tools.Launcher.ViewModels.Home;
