+++ Ksp2Redux.Tools.Launcher.Test/InstallPlanServiceDiskSpaceTest.cs
@@ -1,6 +1,7 @@
 using Ksp2Redux.Tools.Common.Services;
 using Ksp2Redux.Tools.Launcher.Models;
-using Ksp2Redux.Tools.Launcher.Services;
+using Ksp2Redux.Tools.Launcher.Services.Install;
+using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 using Moq;
 using Testably.Abstractions.Testing;
