+++ Ksp2Redux.Tools.Launcher.Test/Ksp2DetectorServiceTest.cs
@@ -1,4 +1,5 @@
-using Ksp2Redux.Tools.Launcher.Services;
+using Ksp2Redux.Tools.Launcher.Services.Install;
+using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 using Moq;
 using Testably.Abstractions.Testing;
