+++ Ksp2Redux.Tools.Launcher.Test/Ksp2InstallRowViewModelValidationTest.cs
@@ -1,5 +1,5 @@
 using Ksp2Redux.Tools.Launcher.Models;
-using Ksp2Redux.Tools.Launcher.Services;
+using Ksp2Redux.Tools.Launcher.Services.Install;
 using Ksp2Redux.Tools.Launcher.ViewModels.Settings;
 using Moq;
 using Testably.Abstractions.Testing;
