+++ Ksp2Redux.Tools.Launcher.Test/LogServiceTest.cs
@@ -1,5 +1,6 @@
 using System.IO.Abstractions;
-using Ksp2Redux.Tools.Launcher.Services;
+using Ksp2Redux.Tools.Launcher.Services.Install;
+using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 using Moq;
 using Testably.Abstractions.Testing;
