+++ Ksp2Redux.Tools.Launcher.Test/NewsServiceTest.cs
@@ -1,5 +1,6 @@
 using CodeHollow.FeedReader;
-using Ksp2Redux.Tools.Launcher.Services;
+using Ksp2Redux.Tools.Launcher.Services.News;
+using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 using Moq;
 
 namespace Ksp2Redux.Tools.Launcher.Test;
