+++ Ksp2Redux.Tools.Launcher.Test/WindowPlacementServiceTest.cs
@@ -1,6 +1,6 @@
 using Avalonia;
 using Ksp2Redux.Tools.Launcher.Models;
-using Ksp2Redux.Tools.Launcher.Services;
+using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 
 namespace Ksp2Redux.Tools.Launcher.Test;
