+++ Ksp2Redux.Tools.Launcher/Services/Feeds/ManifestReleasesFeedProviderService.cs
@@ -8,8 +8,10 @@ using System.Threading.Tasks;
 using Ksp2Redux.Tools.Common.Models;
 using Ksp2Redux.Tools.Launcher.Models;
 using Octokit;
+using Ksp2Redux.Tools.Launcher.Services.Install;
+using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 
-namespace Ksp2Redux.Tools.Launcher.Services;
+namespace Ksp2Redux.Tools.Launcher.Services.Feeds;
 
 public interface IManifestReleasesFeedProviderService
 {
