+++ Ksp2Redux.Tools.Launcher/Services/Infrastructure/LocalStoragePaths.cs
@@ -1,7 +1,7 @@
 using System;
 using System.IO.Abstractions;
 
-namespace Ksp2Redux.Tools.Launcher.Services;
+namespace Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 
 internal static class LocalStoragePaths
 {
