+++ Ksp2Redux.Tools.Launcher/Services/Infrastructure/MessageBoxService.cs
@@ -6,7 +6,7 @@ using MsBox.Avalonia;
 using MsBox.Avalonia.Base;
 using MsBox.Avalonia.Enums;
 
-namespace Ksp2Redux.Tools.Launcher.Services;
+namespace Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 
 public interface IMessageBoxService
 {
