+++ Ksp2Redux.Tools.Launcher/Services/Install/CacheRestoreException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Ksp2Redux.Tools.Launcher.Services;
+namespace Ksp2Redux.Tools.Launcher.Services.Install;
 
 /// <summary>
 /// Thrown by <see cref="CacheService.RecursivelyRestoreCache"/> when the pre-patch snapshot it needs
