+++ Ksp2Redux.Tools.Launcher/Services/Install/Ksp2DetectorService.cs
@@ -2,8 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.IO.Abstractions;
 using System.Text.RegularExpressions;
+using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 
-namespace Ksp2Redux.Tools.Launcher.Services;
+namespace Ksp2Redux.Tools.Launcher.Services.Install;
 
 public interface IKsp2DetectorService
 {
