+++ Ksp2Redux.Tools.Launcher/ViewModels/Settings/Ksp2InstallRowViewModel.cs
@@ -3,7 +3,7 @@ using System.IO.Abstractions;
 using System.Linq;
 using CommunityToolkit.Mvvm.ComponentModel;
 using Ksp2Redux.Tools.Launcher.Models;
-using Ksp2Redux.Tools.Launcher.Services;
+using Ksp2Redux.Tools.Launcher.Services.Install;
 
 namespace Ksp2Redux.Tools.Launcher.ViewModels.Settings;
