+++ NuGet.Config
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <configuration>
-  <!-- Ignore any machine-level feeds so restores are reproducible and only pull from nuget.org. -->
-  <packageSources>
-    <clear />
-    <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
-  </packageSources>
+    <!-- Ignore any machine-level feeds so restores are reproducible and only pull from nuget.org. -->
+    <packageSources>
+        <clear />
+        <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
+    </packageSources>
 </configuration>
